bugfix delete itself

+++ judge/dispatcher.py
@@ -183,7 +183,9 @@ class JudgeDispatcher(DispatcherBase):
                 self.submission.result = JudgeStatus.PARTIALLY_ACCEPTED
         # 如果該次繳交全對：如程式碼與未曾繳交過，則保留，否則拋棄
         if self.submission.result == JudgeStatus.ACCEPTED:
-            for past_submission in Submission.objects.filter(user_id=self.submission.user_id):
+            for past_submission in Submission.objects.filter(user_id=self.submission.user_id, problem=self.problem):
+                if past_submission.id == self.submission.id:
+                    continue
                 if past_submission.code == self.submission.code:
                     past_submission.delete()
                 if past_submission.result != JudgeStatus.ACCEPTED:
@@ -191,10 +193,19 @@ class JudgeDispatcher(DispatcherBase):
             self.submission.save()
         # 如果該次繳交未對：未曾AC，則保留，否則拋棄
         else:
-            for past_submission in Submission.objects.filter(user_id=self.submission.user_id):
+            no_ac = True
+            for past_submission in Submission.objects.filter(user_id=self.submission.user_id, problem=self.problem):
+                if past_submission.id == self.submission.id:
+                    continue
+                #溯及既往
+                if past_submission.result != JudgeStatus.ACCEPTED:
+                    past_submission.delete() 
                 if past_submission.result == JudgeStatus.ACCEPTED:
-                    break
-            self.submission.save()
+                    no_ac = False
+            if no_ac or self.submission.result == JudgeStatus.SYSTEM_ERROR:
+                self.submission.save()
+            else:
+                self.submission.delete()
         if self.contest_id:
             if self.contest.status != ContestStatus.CONTEST_UNDERWAY or \
                     User.objects.get(id=self.submission.user_id).is_contest_admin(self.contest):
